Zero prepend single digit days and months
The yaml template files were updated so that all days and months
are rendered with two digits even if in the range 1 through 9.
This is performed by formatting the integers
The Request class had to be updated in order to parse user input
numeric input from a string form into an integer form. (If a
decimal point is present, the numeric string is parsed to a float
instead.

This enhancement will ensure files lie in correct numerical order in
directories. (Example: before, 10 would appear before 2; now, 02
appears before 10).